add a piechart wrapper on icon and button
