+++ packages/components/src/PieChart/PieChart.component.js
@@ -1,358 +1,51 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { arc } from 'd3-shape';
-import { translate } from 'react-i18next';
-import classnames from 'classnames';
-import I18N_DOMAIN_COMPONENTS from '../constants';
-import getDefaultT from '../translate';
-import Skeleton from '../Skeleton';
-import theme from './PieChart.scss';
-
-export const PIECHART_CONSTANTS = {
-	MIN_SIZE: 20,
-	MAX_SIZE: 50,
-	MAX_PERCENT: 100,
-	BASE_INNER_RADIUS: 6,
-	BASE_OUTER_RADIUS: 9,
-	BASE_PAD_ANGLE: 0.2,
-	INNER_RADIUS_PER_PIXEL: 0.4,
-	OUTER_RADIUS_PER_PIXEL: 0.45,
-	PAD_ANGLE_PER_PIXEL: 0.0033,
-};
-
-const displaySizes = {
-	small: 20,
-	medium: 35,
-	large: 50,
-};
-
-// we need just one instance of this, it's just a generator
-const arcGen = arc();
+import { OverlayTrigger } from 'react-bootstrap';
+import PieChartIcon from './PieChartIcon.component';
+import PieChartButton from './PieChartButton.component';
+import TooltipTrigger from '../TooltipTrigger';
 
 /**
- * This function return the angle for a given percentage
- * @param {number} percentage the percentage to calculate the angle
+ * This function wrap the button with a TooltipTrigger
+ * @param {Element} btn the button element ( may be wrapped by overlay trigger )
+ * @param {boolean} tooltip tell if the tooltip has to be showed
+ * @param {string} label the label to show on the tooltip
+ * @param {string} tooltipPlacement the tooltip placement
  */
-export function getAngle(percentage) {
-	return ((percentage * 2) / 100) * Math.PI;
-}
-
-/**
- * This function return the showed value on the chart
- * @param {array} model the pie chart model
- * @param {index} index current index showed
- */
-function getShowedValue(model, index) {
-	if (!model) {
-		return {};
+export function decorateWithTooltip(tooltip, label, tooltipPlacement, Component, props) {
+	if (!tooltip || !label) {
+		return <Component {...props} />;
 	}
-	return model[index];
-}
-
-/**
- * This function is just a sort
- * @param {object} a an element with a percentageShown to compare
- * @param {object} b an other element with a percentageShown to compare
- */
-function sortElements(a, b) {
-	return b.percentageShown - a.percentageShown;
-}
-
-/**
- * This function check if the type is a number & check with min/max
- * @param {object} props list of component props
- * @param {string} propName current prop name
- * @param {string} componentName component name
- */
-export function propTypeCheckSize(props, propName, componentName) {
-	if (props[propName] == null) {
-		return null;
-	}
-	if (typeof props[propName] !== 'number') {
-		return new Error(
-			`Invalid type of ${propName} supplied to ${componentName} : ${typeof props[
-				propName
-			]}. Validation failed.`,
-		);
-	} else if (
-		props[propName] < PIECHART_CONSTANTS.MIN_SIZE ||
-		props[propName] > PIECHART_CONSTANTS.MAX_SIZE
-	) {
-		return new Error(
-			`Invalid prop ${propName} supplied to ${componentName} : ${
-				props[propName]
-			}. Validation failed.`,
-		);
-	}
-	return null;
-}
-
-/**
- * this function generates chart's empty part
- * @param {array} values the values shown in the graph
- * @param {object} size the current size
- * @param {function} arcGen the arc generator
- * @param {number} minimumPercentage the minimum percentage to be shown
- */
-export function getEmptyPartCircle(values, size, minimumPercentage) {
-	const allPercentages = values.reduce((acc, value) => acc + value.percentageShown, 0);
-	if (allPercentages >= PIECHART_CONSTANTS.MAX_PERCENT - minimumPercentage) {
-		return null;
-	}
-
-	const arcGenerated = arcGen({
-		innerRadius: size.innerRadius,
-		outerRadius: size.outerRadius,
-		padAngle: size.padAngle,
-		startAngle: getAngle(allPercentages),
-		endAngle: Math.PI * 2,
-	});
 	return (
-		<path
-			className={classnames(theme['tc-pie-chart-color-alto'], 'tc-pie-chart-color-alto')}
-			d={arcGenerated}
-			transform={`translate(${size.svgSize / 2},${size.svgSize / 2})`}
-		/>
+		<TooltipTrigger label={label} tooltipPlacement={tooltipPlacement}>
+			<Component {...props} />
+		</TooltipTrigger>
 	);
 }
 
-/**
- * It increase values below minPercentage & decrease values above if some are increased
- * setMinimumPercentage([{percentage: 50}, {percentage: 2}], 5) would result on
- * [{percentage: 50, percentageShown: 47}, {percentage: 2, percentageShown: 5}]
- * @param {*} values the set of values
- * @param {*} minimumPercentage the minimum value we have to show
- * @param {*} amountToSubtract the amount to decrease
- */
-export function distributePercentages(values, minimumPercentage, amountToSubtract) {
-	/**
-	 * This function decrease the percentage shown & the amount to subtract by 1
-	 * @param {object} element the current element
-	 */
-	function decreaseElement(element) {
-		if (amountToSubtract > 0) {
-			// eslint-disable-next-line no-param-reassign
-			element.percentageShown -= 1;
-			// eslint-disable-next-line no-param-reassign
-			amountToSubtract -= 1;
-		}
-	}
-
-	while (amountToSubtract > 0) {
-		const elementsToDecrease = values
-			.filter(value => value.percentageShown > minimumPercentage)
-			.sort(sortElements);
-		if (elementsToDecrease.length > 0) {
-			elementsToDecrease.forEach(decreaseElement);
-		} else {
-			// eslint-disable-next-line no-param-reassign
-			amountToSubtract = 0;
-		}
-	}
-
-	return values;
+export function getPieChartComponent(onClick) {
+	return onClick ? PieChartButton : PieChartIcon;
 }
 
-/**
- * This function sets minimum percentage show with the minimum percentage shown
- * @param {array} values the set of values
- * @param {number} minimumPercentage the minimum value we have to show
- */
-export function setMinimumPercentage(model, minimumPercentage) {
-	if (!model) {
-		return [];
-	}
-
-	let amountToSubtract = 0;
-
-	const valuesMins = model.map(value => {
-		if (value.percentage && value.percentage < minimumPercentage) {
-			amountToSubtract += minimumPercentage - value.percentage;
-			return { ...value, percentageShown: minimumPercentage };
-		}
-		return { ...value, percentageShown: value.percentage };
-	});
-
-	return distributePercentages(valuesMins, minimumPercentage, amountToSubtract);
-}
-
-/**
- * This function return the sum of the percentages below the given index
- * @param {array} values the values to get when we have to start
- * @param {number} index the current index
- */
-export function getPercentageToIndex(values, index) {
-	return values.reduce((acc, value, i) => {
-		if (i < index) {
-			return acc + value.percentageShown;
-		}
-		return acc;
-	}, 0);
-}
-
-/**
- * This function return the label or nothing if the label is npt passed or hidden
- * @param {boolean} hideLabel tell if the label has to be hidden or not
- * @param {number} labelValue the label value ( percentage )
- * @param {function} t translate function
- */
-function getLabel(hideLabel, labelValue, t) {
-	if (!hideLabel && labelValue.percentage != null) {
-		return t('PIE_CHART_PERCENTAGE', {
-			defaultValue: '{{percentage}}%',
-			percentage: labelValue.percentage,
-		});
-	}
-	return '';
-}
-
-/**
- * This function generate a part of circle for a value
- * @param {object} value the value to generate
- * @param {number} index the current index
- * @param {array} values the values to get when we have to start
- * @param {object} size the current graph size
- */
-export function getCircle(value, index, values, size) {
-	const percentagesDone = getPercentageToIndex(values, index);
-
-	const arcGenerated = arcGen({
-		innerRadius: size.innerRadius,
-		outerRadius: size.outerRadius,
-		padAngle: size.padAngle,
-		startAngle: getAngle(percentagesDone),
-		endAngle: getAngle(percentagesDone + value.percentageShown),
-	});
-
-	return (
-		<path
-			key={index}
-			d={arcGenerated}
-			className={classnames(
-				theme[`tc-pie-chart-color-${value.color}`],
-				`tc-pie-chart-color-${value.color}`,
-			)}
-			transform={`translate(${size.svgSize / 2},${size.svgSize / 2})`}
-		/>
+export default function PieChart({ tooltip, tooltipPlacement, ...props }) {
+	return decorateWithTooltip(
+		tooltip,
+		props.label,
+		tooltipPlacement,
+		getPieChartComponent(props.onClick),
+		props,
 	);
 }
 
-/**
- * This function return useful stuff to build the graph of the loader
- * @param {number} size the size in px of the graph
- * @param {string} display the chosen display if given
- */
-export function getDisplaySize(size, display) {
-	let currentSize = size;
-	if (!currentSize && display) {
-		currentSize = displaySizes[display];
-	}
-
-	const pixelNumber = currentSize - PIECHART_CONSTANTS.MIN_SIZE;
-	return {
-		svgSize: currentSize,
-		innerRadius: parseInt(
-			PIECHART_CONSTANTS.BASE_INNER_RADIUS +
-				PIECHART_CONSTANTS.INNER_RADIUS_PER_PIXEL * pixelNumber,
-			10,
-		),
-		outerRadius: parseInt(
-			PIECHART_CONSTANTS.BASE_OUTER_RADIUS +
-				PIECHART_CONSTANTS.OUTER_RADIUS_PER_PIXEL * pixelNumber,
-			10,
-		),
-		padAngle:
-			PIECHART_CONSTANTS.BASE_PAD_ANGLE - PIECHART_CONSTANTS.PAD_ANGLE_PER_PIXEL * pixelNumber,
-	};
-}
-
-export function PieChartComponent({
-	display,
-	hideLabel,
-	labelIndex,
-	loading,
-	minimumPercentage,
-	model,
-	size,
-	t,
-}) {
-	const sizeObject = getDisplaySize(size, display);
-	if (loading) {
-		return (
-			<span className={classnames(theme['tc-pie-chart-loading'], 'tc-pie-chart-loading')}>
-				<Skeleton
-					type={Skeleton.TYPES.circle}
-					width={sizeObject.svgSize}
-					height={sizeObject.svgSize}
-					className={classnames(
-						theme['tc-pie-chart-loading-circle'],
-						'tc-pie-chart-loading-circle',
-					)}
-				/>
-				{!hideLabel && <Skeleton type={Skeleton.TYPES.text} size={Skeleton.SIZES.small} />}
-			</span>
-		);
-	}
-	const labelValue = getShowedValue(model, labelIndex);
-	const preparedValues = setMinimumPercentage(model, minimumPercentage);
-	return (
-		<span className={classnames(theme['tc-pie-chart'], 'tc-pie-chart')}>
-			<svg
-				width={sizeObject.svgSize}
-				height={sizeObject.svgSize}
-				className={classnames(theme['tc-pie-chart-graph'], 'tc-pie-chart-graph')}
-				style={{ width: sizeObject.svgSize, height: sizeObject.svgSize }}
-			>
-				{preparedValues.map((value, index) => getCircle(value, index, preparedValues, sizeObject))}
-				{getEmptyPartCircle(preparedValues, sizeObject, minimumPercentage)}
-			</svg>
-			<div
-				className={classnames(
-					theme['tc-pie-chart-label'],
-					'tc-pie-chart-label',
-					theme[`tc-pie-chart-color-${labelValue.color}`],
-					`tc-pie-chart-color-${labelValue.color}`,
-				)}
-			>
-				{getLabel(hideLabel, labelValue, t)}
-			</div>
-		</span>
-	);
-}
-
-export const pieChartPropTypes = {
-	display: PropTypes.oneOf(['small', 'medium', 'large']),
-	hideLabel: PropTypes.bool,
-	labelIndex: PropTypes.number,
-	loading: PropTypes.bool,
-	minimumPercentage: PropTypes.number,
-	model: PropTypes.arrayOf(
-		PropTypes.shape({
-			color: PropTypes.oneOf([
-				'rio-grande',
-				'chestnut-rose',
-				'lightning-yellow',
-				'dove-gray',
-				'silver-chalice',
-			]),
-			percentage: PropTypes.number.isRequired,
-		}).isRequired,
-	),
-	size: propTypeCheckSize,
+PieChart.propTypes = {
+	label: PropTypes.string,
+	onClick: PropTypes.func,
+	tooltip: PropTypes.bool,
+	tooltipPlacement: OverlayTrigger.propTypes.placement,
 };
 
-PieChartComponent.propTypes = {
-	...pieChartPropTypes,
-	t: PropTypes.func,
+PieChart.defaultProps = {
+	tooltipPlacement: 'top',
 };
 
-PieChartComponent.defaultProps = {
-	labelIndex: 0,
-	minimumPercentage: 5,
-	display: 'small',
-	t: getDefaultT(),
-};
-
-PieChartComponent.displayName = 'PieChart';
-
-export default translate(I18N_DOMAIN_COMPONENTS)(PieChartComponent);
+PieChart.displayName = 'PieChart';
