+++ packages/components/src/PieChart/index.js
@@ -1,4 +1,5 @@
 import PieChartButton from './PieChartButton.component';
+import PieChartIcon from './PieChartIcon.component';
 import PieChart from './PieChart.component';
 
-export { PieChartButton, PieChart };
+export { PieChart as default, PieChartButton, PieChartIcon };
